requirements.txt
django==1.10
pyOpenSSL
mysql==0.0.1
MySQL-python==1.2.5